Remove usage of sys/sysctl.h (deprecated) in non-BSD builds
The header file sys/sysctl.h is deprecated in Linux from some time.
This means that a warning is now issued that may break the build if
error on warnings is enabled. In the future, as the header gets
removed, this will be directly an error.
It seems that we still need for now this in BSD as sysctl is used
to get the number of processors, so confine the usage of such header
to BSD builds via ifdef.

Signed-off-by: Federico Pellegrin <fede@evolware.org>